Fixed delta calc for real

+++ src/main/java/org/spout/api/gui/ScreenStack.java
@@ -33,7 +33,7 @@ import org.spout.api.signal.SignalSubscriberObject;
 import org.spout.api.tickable.Tickable;
 
 public class ScreenStack extends SignalSubscriberObject implements Tickable, Runnable {
-	private float lastTick = 0;
+	private long lastTick = 0;
 	LinkedList<Screen> screens = new LinkedList<Screen>();
 	LinkedList<Screen> visibleScreens = null;
 	
@@ -116,9 +116,9 @@ public class ScreenStack extends SignalSubscriberObject implements Tickable, Run
 	@Override
 	public void run() {
 		float delta = 50f;
-		float current = System.currentTimeMillis();
+		long current = System.currentTimeMillis();
 		if (lastTick != 0) {
-			delta = current - lastTick;
+			delta = (float) (current - lastTick);
 		}
 		lastTick = current;
 		tick(delta);
